+++ app/src/main/java/com/example/dancognitionapp/bart/data/BartUiState.kt
@@ -1,6 +1,5 @@
 package com.example.dancognitionapp.bart.data
 
-import com.example.dancognitionapp.bart.model.Balloon
 import java.util.LinkedList
 
 data class BartUiState(
